+++ tools/tests/tester/StoryValidator.test.ts
@@ -31,10 +31,11 @@ describe('StoryValidator', () => {
     const evaluation = validate('tools/tests/tester/fixtures/invalid_story.yaml')
     expect(evaluation?.result).toBe('ERROR')
     expect(evaluation?.message).toBe("Invalid Story: " +
-      "data/epilogues/0 contains unsupported properties: response --- " +
       "data/chapters/0 MUST contain the missing properties: method --- " +
       "data/chapters/1/method MUST be equal to one of the allowed values: GET, PUT, POST, DELETE, PATCH, HEAD, OPTIONS --- " +
-      "data/chapters/1/method must be array --- data/chapters/1/method must match exactly one schema in oneOf")
+      "data/chapters/1/method must be array --- " +
+      "data/chapters/1/method must match exactly one schema in oneOf"
+    )
   })
 
   test('invalid description', () => {